make hard wrap behavior configurable in content settings
Hard wraps turn line breaks into <br>, which is something that
SilverBullet content directories need for proper rendering,
but which breaks rendering hugo content directories
(or even the README file in the bull repository!)

Content settings are configured in a TOML file in the _bull directory.
This directory will never be served by bull (instead, bull’s own
handlers are served under the _bull/ prefix) and is clearly identifiable
as belonging to bull, so this arrangement seems like a good choice.