bull.Customization: remove AfterPageParse
This API was not useful: without the source, you cannot even obtain the context
of *ast.Text nodes, for example (which seems like a pretty basic operation).

Furthermore, AST nodes cannot be modified: the Goldmark AST is not a copy of the
input data, but rather uses indexes to point into the input data.

Hence, for customization, it is typically better to either supply a custom
parser (which adds nodes to the AST as desired), a custom renderer (which takes
an existing AST and renders it in a different way) or a mixture of both.

I’ll add an API to supply Goldmark extensions in a follow-up commit instead.