implement interactive task list functionality (tick/un-tick tasks)
This functionality is enabled by default when the editor is enabled,
but can be turned off through the content setting interactive_task_list.

Each checkbox is identified by its line number.
Ticking/Un-ticking a checkbox results in a request to the server,
which adds/removes the 'x' inside [ ] and saves the file.